attachments/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pillow==10.1.0
pytesseract==0.3.10
python-dotenv==1.0.0
